使用 NodeInfoEditorDirector 管理节点编辑窗口

+++ srcs/cpp/MagicRenderView/srcs/node/node/imp/beginNodes/startNode.h
@@ -8,7 +8,7 @@ class StartNode : public BeginNode {
 	Q_OBJECT;
 protected:
 	bool readUpdate;
-	int* var;
+	size_t *var;
 	BeginOutputPort *beginOutputPort;
 public:
 	StartNode( const QString &node_name );
@@ -20,7 +20,7 @@ public:
 	bool readyNodeRunData( ) override;
 	bool fillInputPortCall( const QDateTime &ndoe_run_start_data_time, std::vector< Node * > &result_need_run_ref_node_vector ) override;
 	bool fillNodeCall( const QDateTime &ndoe_run_start_data_time ) override;
-	bool fillOutputPortCall( std::vector<Node *> &result_next_run_advise_node_vector, const QDateTime &ndoe_run_start_data_time ) override;
+	bool fillOutputPortCall( std::vector< Node * > &result_next_run_advise_node_vector, const QDateTime &ndoe_run_start_data_time ) override;
 public:
 	Def_Extend_NodeTypeName( Node::tr( "起始/开始" ) );
 };
